Refactor datasets test (#817)
